requirements.txt
pandas==0.25.2
numpy==1.17.3
torch==1.3.0
pytorch-ignite==0.3.0
tqdm==4.36.1
nltk==3.5
tableprint==0.8.0
fire==0.3.0
zhon==1.1.5
scikit_learn==0.21.3
PyYAML==5.1.2
gensim==3.8.3
